ws1: Mention using HTML Notebook for full results
With PDF documents, displaying large data frames fully can be annoying
since this means more space is taken up. This is not the case with HTML
Notebooks, since the data frame can be paginated, thus saving on
vertical and horizontal space.

Some questions in this worksheet mention displaying data frames, so
mention HTML Notebooks for each instance as an alternative to fully
 display the results.