ws4: q2: p3: Correct answer regarding CLT
Previously, it was said that the sample sizes for each n-bedroom group
were not large enough for CLT to apply. From lecture, a sample size of
around 10 or above actually suffices.

Here, the problem lies with the skewness and outliers in the 3-bedroom
group. In particular, these two factors may prevent a sample size of 23
from allowing CLT to kick in, so mention this instead.