ws4: q2: p3: Draw plot of bootstrapped mean difference
We are interested in whether a two-sample t-test applies here, so what
really matters is the whether the mean difference of the two groups is
normally distributed. This also allows for just one histogram instead of
two.

Explain the rationale for examining the mean difference for normality
and draw the corresponding histogram.