ws5: q2-3: Ungroup row-wise later in simulations
ungroup() was originally applied to the p-value tables in simulations
for Questions 2 and 3 (to undo rowwise()'s row-wise grouping), but it's
only necessary when using summarize() to count the total simulation
runs in column n.

Move the ungroup() to the relevant summarize() line, which also hides it
from the final PDF document (due to some lines being deliberately hidden
helper variable lines).