Always break lines before horizontal rule
When using \rule in LaTeX to create a horizontal rule separating blocks
of text, the \rule needs to be on its own line to ensure that it doesn't
create a horizontal rule on the same line as the last segment of text.

If the previous line extends all the way to the end, then no action is
needed. However, if it does not, action must be taken. This can be done
with \par, but doing so when not needed occurs in massive vertical
space. Breaking the line with \\ in all cases results in acceptable
vertical spacing, so always use break lines with \\ before horizontal
rules for consistency.